break out parseMessage
