+++ packages/grafana-ui/src/components/Logs/LogRowMessageParsed.tsx
@@ -1,16 +1,10 @@
 import React, { PureComponent } from 'react';
-import memoizeOne from 'memoize-one';
-import { Field, getParser, LinkModel, LogRowModel } from '@grafana/data';
+import { LogRowModel } from '@grafana/data';
 
 import { Themeable } from '../../types/theme';
 import { withTheme } from '../../themes/index';
 
-type FieldDef = {
-  key: string;
-  value: string;
-  links?: Array<LinkModel<Field>>;
-  fieldIndex?: number;
-};
+import { parseMessage } from './logParser';
 
 export interface Props extends Themeable {
   row: LogRowModel;
@@ -18,27 +12,9 @@ export interface Props extends Themeable {
 }
 
 class UnThemedLogRowMessageParsed extends PureComponent<Props> {
-  getParser = memoizeOne(getParser);
-
-  parseMessage = memoizeOne((rowEntry): FieldDef[] => {
-    const parser = this.getParser(rowEntry);
-    if (!parser) {
-      return [];
-    }
-    // Use parser to highlight detected fields
-    const parsedFields = parser.getFields(rowEntry);
-    const fields = parsedFields.map(field => {
-      const key = parser.getLabelFromField(field);
-      const value = parser.getValueFromField(field);
-      return { key, value };
-    });
-
-    return fields;
-  });
-
   render() {
     const { row, showParsedFields } = this.props;
-    const fields = this.parseMessage(row.entry);
+    const fields = parseMessage(row.entry);
 
     return (
       <td>
